пока вернула в один файл компоненты
