+++ src/widgets/products/TableProduct.module.css
